// src/types/errors.ts
export enum ERRORS {
    name = 'Первая буква должна быть заглавной, без пробелов и без цифр, из спецсимволов допустим только дефис',
    login = 'От 3 до 20 символов латиницы может содержать цифры, без пробелов, допустимы дефис и нижнее подчёркивание',
    email = 'Латиница, может включать цифры и спецсимволы вроде дефиса, обязательно @ и точка после',
    password = 'От 8 до 40 символов, обязательно хотя бы одна заглавная буква и цифра',
    phone = 'От 10 до 15 символов, состоит из цифр, может начинается с плюса',
    message = 'Не должно быть пустым',
    form = 'Пароли не должны совпадать'
}